+++ src/lib.rs
@@ -49,7 +49,6 @@
 //! ```
 
 #![cfg_attr(not(test), no_std)]
-#[macro_use]
 extern crate alloc;
 
 pub mod accept;